[flang][NFC] Move OpenMP related passes into a separate directory
